Put Alpha project version 1.0.0

+++ controllers/userController.js
@@ -59,15 +59,50 @@ exports.post_user = asyncHandler(async (req, res, next) => {
     const hashedPassword = await genPass(req.body.password);
     const userId = uuidv4();
     const defaultPicturePath = process.env.DEFAULT_PICTURE_PATH;
-    await req.prisma.user.create({
-        data: {
-            id: userId,
-            email: req.body.email,
-            name: req.body.name,
-            password: hashedPassword,
-            picture: defaultPicturePath,
+    let response = await fetch(
+        'https://mosquitto-api.onrender.com/api/client/add/record',
+        {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({
+                username: userId,
+                password: '0000',
+                clientid: userId,
+                textname: req.body.name || '',
+                textdescription:
+                    `user Client with, Email:${req.body.email}` || '',
+                roles: [{ rolename: 'lwt', priority: 0 }],
+            }),
         },
-    });
+    );
+    if (response.ok) {
+        response = await fetch(
+            'https://mosquitto-api.onrender.com/api/client/enable/record',
+            {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify({
+                    username: userId,
+                    enable: true,
+                }),
+            },
+        );
+    }
+    if (response.ok) {
+        await req.prisma.user.create({
+            data: {
+                id: userId,
+                email: req.body.email,
+                name: req.body.name,
+                password: hashedPassword,
+                picture: defaultPicturePath,
+            },
+        });
+    }
     res.status(201).json({
         success: true,
         status: 201,
@@ -164,22 +199,44 @@ exports.delete_user = asyncHandler(async (req, res, next) => {
             id: req.body.id,
         },
     });
-    await req.prisma.user.delete({
-        where: {
-            id: req.body.id,
-        },
-    });
-    if (result.picture !== process.env.DEFAULT_PICTURE_PATH) {
-        const filePath = path.resolve(
-            __dirname,
-            `${process.env.GET_PICTURE_PATH + result.picture}`,
-        );
-        await fs.access(filePath);
-        await fs.unlink(filePath);
+    if (result) {
+        await req.prisma.user.delete({
+            where: {
+                id: req.body.id,
+            },
+        });
+        if (result.picture !== process.env.DEFAULT_PICTURE_PATH) {
+            const filePath = path.resolve(
+                __dirname,
+                `${process.env.GET_PICTURE_PATH + result.picture}`,
+            );
+            await fs.access(filePath);
+            await fs.unlink(filePath);
+        }
     }
-    res.status(200).json({
-        success: true,
-        status: 200,
-        message: 'User deleted successfully',
-    });
+    const response = await fetch(
+        'https://mosquitto-api.onrender.com/api/client/delete/record',
+        {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({
+                username: req.body.id,
+            }),
+        },
+    );
+    if (response.ok) {
+        res.status(200).json({
+            success: true,
+            status: 200,
+            message: 'User deleted successfully',
+        });
+    } else {
+        res.status(200).json({
+            success: false,
+            status: 500,
+            message: 'Error deleting user from MQTT',
+        });
+    }
 });
